\name{ThreeGroups-package}
\alias{ThreeGroups-package}
\alias{ThreeGroups}
\docType{package}
\title{ML Estimator for Baseline-Placebo-Treatment (Three-Group) Experiments}
\description{This package implements the Maximum Likelihood estimator for baseline, placebo, and treatment groups (three-group) experiments with non-compliance proposed by Gerber, Green, Kaplan, and Kern (2010).}
\details{
\tabular{ll}{
Package: \tab ThreeGroups\cr
Type: \tab Package\cr
Version: \tab 0.21\cr
Date: \tab 2015-09-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}


\author{Holger L. Kern, \email{hkern@fsu.edu}, \url{http://www.holgerkern.com}}


\references{Alan S. Gerber, Donald P. Green, Edward H. Kaplan, and Holger L. Kern. 2010. ``Baseline, placebo, and treatment: Efficient estimation for three-group experiments.'' \emph{Political Analysis} 18 (3): 297-315.}
