\name{ThreeGroups-package}
\alias{ThreeGroups-package}
\alias{ThreeGroups}
\docType{package}
\title{ML Estimator for Baseline-Placebo-Treatment (Three-group) designs}
\description{This package implements the Maximum Likelihood estimator for three-group designs proposed by Gerber, Green, Kaplan, and Kern (2010).}
\details{
\tabular{ll}{
Package: \tab ThreeGroups\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-06-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}


\author{Holger L. Kern, \email{holger.kern@yale.edu}, \url{http://pantheon.yale.edu/~hlk8/}}


\references{Alan Gerber, Donald Green, Edward H. Kaplan, and Holger L. Kern. 2010. ``Baseline, placebo, and treatment: Efficient estimation for three-group experiments.'' \emph{Political Analysis}.}
